backend/requirements.txt
Flask
flask-pymongo
gunicorn
pillow
python-dotenv
captcha